# api/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'College Attendance API'
  description: 'Versioned REST API for marking attendance, managing students/subjects, and generating detention & dashboard reports.'
  version: 1.0.0
servers:
  -
    url: /api/v1
    description: 'Default API server'
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: 'Log in with email/username and password, receiving a JWT access + refresh token pair'
      operationId: 985cd5203e2d0a820e113f2f4fdb2847
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  description: 'Email or username'
                  type: string
                  example: admin@college.edu
                password:
                  type: string
                  format: password
              type: object
      responses:
        200:
          description: 'Login successful'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      access_token:
                        type: string
                      refresh_token:
                        type: string
                      token_type:
                        type: string
                        example: bearer
                      expires_in:
                        description: 'Seconds until access token expiry'
                        type: integer
                      user:
                        type: object
                    type: object
                type: object
        401:
          description: 'Invalid credentials'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        423:
          description: 'Account locked due to repeated failed attempts'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: 'Exchange a valid refresh token for a new access + refresh token pair (rotates the refresh token)'
      operationId: 03c7bb98658eb98880114621321bd8d8
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
              type: object
      responses:
        200:
          description: 'New token pair issued'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      access_token:
                        type: string
                      refresh_token:
                        type: string
                      token_type:
                        type: string
                        example: bearer
                      expires_in:
                        type: integer
                    type: object
                type: object
        401:
          description: 'Refresh token invalid, expired, or user not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: 'Log out — denylists the current access token and (optionally) revokes the refresh token'
      operationId: a05022a34eef599e04fb0a5254375b84
      requestBody:
        required: false
        content:
          application/json:
            schema:
              properties:
                refresh_token:
                  type: string
                  nullable: true
              type: object
      responses:
        200:
          description: 'Logged out'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      message:
                        type: string
                    type: object
                type: object
      security:
        -
          bearerAuth: []
  /auth/me:
    get:
      tags:
        - Auth
      summary: 'Get the currently authenticated user'
      operationId: e339e0e74860c0647fd1d41bbbbc13b2
      responses:
        200:
          description: 'Current user'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    properties:
                      user_id:
                        type: integer
                      username:
                        type: string
                      email:
                        type: string
                      full_name:
                        type: string
                      role:
                        type: string
                      faculty_id:
                        type: integer
                        nullable: true
                      department:
                        type: string
                        nullable: true
                    type: object
                type: object
        401:
          description: 'Invalid or expired token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    ApiError:
      properties:
        success:
          type: boolean
          example: false
        error:
          properties:
            code:
              type: string
              example: VALIDATION_FAILED
            message:
              type: string
              example: 'The given data was invalid.'
            details:
              type: object
              nullable: true
          type: object
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Enter the JWT access token returned by /api/v1/auth/login'
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []